add base detector node
